git-svn-id: http://svn.code.sf.net/p/cgns/code/cgns/trunk@166 59ee3124-ab72-4556-8a47-5b88dfbcc004

+++ src/cgnstools/tkogl/feedback.c
@@ -192,7 +192,7 @@ Feedback (Tcl_Interp *interp, int argc, char* argv [])
    }
 
    /* Parse 'type' */
-   len = strlen (argv [3]);
+   len = (int)strlen (argv [3]);
    if (strncasecmp (argv [3], "2d", len) == 0) {
       type = GL_2D;
       proc = PrintVtx2D;
